+++ src/pages/CreateMessagePage.jsx
@@ -5,7 +5,7 @@ import TextInputSection from '../components/TextInputSection';
 import ProfileImgInputSection from '../components/ProfileImgInputSection';
 import TextareaInputSection from '../components/TextareaInputSection';
 import ToggleDownSection from '../components/ToggleDownSection';
-import { createMessage } from '../Api/messageApi';
+import { createMessage } from '../Api/api';
 import Header from '../components/Header';
 
 const CreateMessagePage = () => {
